Start gluing our system to Brig.
There are a couple of things we do to sit within Brig, the big one is where we add our command context to the String Reader. We do this to allow us to pass the context to parsers without mixins and further headaches with trying to make things painless for mods that use Brig directly.

This isn't complete, but it's a good way to what we're expecting. Registrars and parsers will come last.